+++ src/core/shared/enums/InventorySlotType.ts
@@ -1,0 +1,6 @@
+export enum SLOT_TYPE {
+    INVENTORY = 'i-',
+    TOOLBAR = 't-',
+    EQUIPMENT = 'e-',
+    GROUND = 'g-',
+}