+++ documentation/changelog.md
@@ -81,6 +81,7 @@ description: List of all the changes and current versions.
 + Added Conversion from legacy Inventory
 + Updated Inventory, Equipment, etc. functions.
 + Added World Notifications (3D Help Text) Functions
++ Added Custom Inventory Rule Support for All Inventory Action(s)
 ```
 
 ## 2.0.3